requirements.txt
aniso8601==8.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==1.1.0
jaraco.classes==3.3.1
Jinja2==2.11.2
keyring==24.3.0
MarkupSafe==1.1.1
more-itertools==10.2.0
numpy==1.26.4
pandas==2.2.0
pycurl==7.45.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2020.1
pyzmq==25.1.2
requests==2.31.0
six==1.15.0
SQLAlchemy==1.3.18
tzdata==2023.4
urllib3==2.2.0
Werkzeug==1.0.1
zipp==3.17.0
